Continued detekt style updates
- Formatting (with Ctrl+Alt+O, Ctrl+Alt+L)
- Naming magic numbers
- Reducing too-long lines
- Detekt now ignores composable function naming as this follows different style rules
- e.t.c.

+++ app/src/main/java/com/turbofan3360/openeq/ui/utils/MathUtils.kt
@@ -4,7 +4,7 @@ import androidx.compose.ui.geometry.Offset
 
 fun roundOneDP(floatValue: Float): Float {
     // Utility to round floats to one decimal place
-    return ((floatValue * 10).toInt()).toFloat() / 10
+    return (floatValue * 10f).toInt() / 10f
 }
 
 fun generateSplineControlPoint(
@@ -14,8 +14,8 @@ fun generateSplineControlPoint(
     nextNextPoint: Offset
 ): Pair<Offset, Offset> {
     // Uses the Catmull-Rom algorithm to generate control points for a spline
-    val controlPoint1 = point + (nextPoint-prevPoint)/6f
-    val controlPoint2 = nextPoint - (nextNextPoint-point)/6f
+    val controlPoint1 = point + (nextPoint - prevPoint) / 6f
+    val controlPoint2 = nextPoint - (nextNextPoint - point) / 6f
 
     return Pair(controlPoint1, controlPoint2)
 }
